requirements.txt
nonebot2
nonebot-adapters-cqhttp
selenium
matplotlib
numpy
pillow
requests
